+++ pkgs/usb_keyboard_scanner/example/main.dart
@@ -1,10 +1,6 @@
 import 'package:combined_barcode_scanner/combined_barcode_scanner.dart';
-import 'package:combined_barcode_scanner_blue_bird/combined_barcode_scanner_blue_bird.dart';
-import 'package:combined_barcode_scanner_fast/combined_barcode_scanner_fast.dart';
-import 'package:combined_barcode_scanner_honeywell/combined_barcode_scanner_honeywell.dart';
-import 'package:combined_barcode_scanner_unitech/combined_barcode_scanner_unitech.dart';
 import 'package:combined_barcode_scanner_usb_keyboard/combined_barcode_scanner_usb_keyboard.dart';
-import 'package:combined_barcode_scanner_zebra/combined_barcode_scanner_zebra.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -60,25 +56,14 @@ class _MyHomePageState extends State<MyHomePage> {
       body: BarcodeScannerWidget(
         controller: _controller,
         onScan: (code) {
-          print("GOT BARCODE =========== ${code.code}");
+          if (kDebugMode) {
+            print("GOT BARCODE =========== ${code.code}");
+          }
         },
         configuration: const ScannerConfiguration(
-          enableFormats: {BarcodeFormat.qr, BarcodeFormat.code128},
-          cameraConfiguration: CameraConfiguration(
-            frameRate: 30,
-            mode: BarcodeDetectionMode.continuous,
-            resolution: CameraResolution.medium,
-            type: CameraType.back,
-          ),
+          enableFormats: {}, // enableFormats unfortunately does not work with the UsbKeyboardScanner, this configuration is managed by the device itself
         ),
-        scanners: [
-          FastBarcodeScanner(),
-          HoneywellBarcodeScanner(),
-          UnitechBarcodeScanner(),
-          BlueBirdBarcodeScanner(),
-          ZebraBarcodeScanner('my_profile'),
-          UsbKeyboardScanner(),
-        ],
+        scanners: [UsbKeyboardScanner()],
       ),
     );
   }