+++ pkgs/usb_keyboard_scanner/lib/combined_barcode_scanner_usb_keyboard.dart
@@ -1,0 +1,3 @@
+library combined_barcode_scanner_usb_keyboard;
+
+export 'src/usb_keyboard_scanner.dart';