+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.1-dev.2
+
+* Add ScannerType source to BarcodeScanResult
+
 ## 0.0.1-dev.1
 
 * Make format in result optional